Move save hook into its own file
useNestedLocalStorage
