+++ src/pages/Game.jsx
@@ -31,36 +31,6 @@ function useLocalStorageRef(key, { defaultValue, replacer }) {
   return ref;
 }
 
-// Perhaps I could write a function that allows for arbitrary amounts of nesting...
-function useNestedLocalStorage({
-  localStorageKey,
-  objectKey,
-  defaultValue,
-  replacer,
-}) {
-  const [val, setVal] = useState(
-    JSON.parse(localStorage.getItem(localStorageKey))?.[objectKey] ||
-      defaultValue
-  );
-
-  useEffect(() => {
-    const prev = JSON.parse(localStorage.getItem(localStorageKey));
-
-    localStorage.setItem(
-      localStorageKey,
-      JSON.stringify(
-        {
-          ...prev,
-          [objectKey]: val,
-        },
-        replacer
-      )
-    );
-  }, [val]);
-
-  return [val, setVal];
-}
-
 export default function Game({ fetchTrivia }) {
   const [turn, setTurn] = useState(TEAM_ONE);
